Bug fix: Fan import should not use encoded_size
Calls to Fan.v and Fan.import should use the same value for ~hash_size.
The bug occured in the Fan.import called by an outdated readonly instance that wants to synchronize with the readwrite instance. An incorrect
hash_size in the fan results in having only some parts of the index accesible by the readonly instance.