Refactor benchs
Replaced list with array: this decreases the execution time of random
read and write to match the performance of bench/main.ml. However, after
sorting the array, the execution time increases significantly, and it
might not be due to index. Therefore the benchs using an ordered array
should not be compared with the ones on random.
Other changes:
removed file size; print stats for lmdb (in case we want to use it);
print nb of operations / seconds.