IO: remove the cache of buffers
We now have a cache of files, so this and adds another layer of
caching with all its unecessary complexity.

Also it's an issue to share these buffer when the underlying
file is being renamed (e.g. merge -> data). This is causing
issues when a concurrent merge and clear are touching this
shared buffer (when obviously they shouldn't, as they are not
for the same file!)