Use Semaphore compatibility library
Our usage of the `Mutex` module in the UNIX implementation of Index
previously relied upon undefined behaviours of `glibc`, which caused
Index to be incompatible with FreeBSD. In particular, we were relying on
being able to release a mutex from a thread that did not itself take the
mutex (in order to implement the merge lock).

This commit replaces the usage of `Mutex` with a dependency on the
`semaphore-compat` library, which provides a shim of the `Semaphore`
module to be released in OCaml 4.12. In particular, the mutual-exclusion
semantics that we require are provided by the binary semaphores in this
library.