Fix clear
Co-authored-by: Gabriel Belouze <gabriel.belouze@ens.psl.eu>

The fix consists in exploiting the fact that the readonly’s fd does not
change even if we rename the file. Then clear: (1) renames file in a
temporary file; (2) opens new file with the new generation; (3) set new
generation in temporary file; (4) closes and removes the temporary file.
This ensures that the readonly only sees a generation change when the
new file is already available on disk.