Crash-consistency bug fix
```ocaml
(* Entry *)
let encode' key value f =
  f (K.encode key);
  (* Flush possible. Crash-consistency bug *)
  f (V.encode value)
```ocaml

Called like this:

```ocaml
Entry.encode' key value (IO.append t)
```ocaml

```ocaml
let append t buf =
  Buffer.add_string t.buf buf;
  flush t (* if offset is high enough *)
```ocaml

This means that the encoding function takes a writing function `f` and applies it to the encoding. If this function `f` flushes a flush may occur and if a crash happens at the same time the flushed encoding will be incomplete (the key will have been written but not the value)

The easiest way to prevent it is to call `f` on the concatenation of the two encoding since this won't induce any API change.