Simplifications on top of #288
We don't need to create a temporary file, it's fine to just remove
the initial file while holding an open file descriptor on the resource.

We also can directly remove the files which are not needed after a clear:
only the log file is needed as it contains the current generation number.