index: recover from previous crashed merges before starting a new one
When the merge thread is failing (for instance because of oom) the exception is
not propagated to the main thread which might continue its life. In that case,
the main thread can decide to start a new merge. Currently doing this will
ignore (and overwrite) any existing [log_async] file, which will cause data
loss.

To avoid this, we now always try to recover from potentially failed merges
before starting a new merge: to do so we transfer the contents of log_async
into log, as the normal startup process does.