index: keep the `Log_file` hashset in sorted order
Previously, the hash-set bucket used to store binding offsets was
determined by the bottom bits of the key's hash. By switching to use the
_top_ bits instead, the hashset keeps entries roughly in order (with
only the entries within each bucket being relatively out-of-order).

This means that we don't need to load all the bindings into an array in
order to sort them for merging, reducing the peak memory usage of Index.