index: use an LRU to cache reads
Following the recent change to the log-file implementation, successful
reads of the index must always read from disk (because even log file
entries are not kept entirely in memory). This traded some speed
performance in favour of much lower memory usage.

For workflows with temporal locality, it's possible to recover some of
this speed by adding a relatively small LRU to cache `find`s from disk.
This is implemented by this commit.