index: fix bug with LRU and `filter`
The `Index.filter` operation may remove bindings that are cached in the
LRU, in which case the read-write instance should be sure to remove them
from the LRU as well. This fix does the simple thing of just clearing
the LRU whenever any `filter` is passed (regardless of which bindings
have been filtered, and whether or not they appear in the LRU).